+++ src/clients/decorators/wallet.test.ts
@@ -17,6 +17,7 @@ test('default', async () => {
       "requestPermissions": [Function],
       "sendTransaction": [Function],
       "signMessage": [Function],
+      "signTypedData": [Function],
       "switchChain": [Function],
       "watchAsset": [Function],
       "writeContract": [Function],
@@ -72,6 +73,63 @@ describe('smoke test', () => {
       await walletClient.signMessage({
         account: getAccount(accounts[0].address),
         data: '0xdeadbeaf',
+      }),
+    ).toBeDefined()
+  })
+
+  test('signTypedData', async () => {
+    expect(
+      await walletClient.signTypedData({
+        account: getAccount(accounts[0].address),
+        domain: {
+          name: 'Ether Mail',
+          version: '1',
+          chainId: 1,
+          verifyingContract: '0x0000000000000000000000000000000000000000',
+        },
+        types: {
+          Name: [
+            { name: 'first', type: 'string' },
+            { name: 'last', type: 'string' },
+          ],
+          Person: [
+            { name: 'name', type: 'Name' },
+            { name: 'wallet', type: 'address' },
+            { name: 'favoriteColors', type: 'string[3]' },
+            { name: 'age', type: 'uint8' },
+            { name: 'isCool', type: 'bool' },
+          ],
+          Mail: [
+            { name: 'timestamp', type: 'uint256' },
+            { name: 'from', type: 'Person' },
+            { name: 'to', type: 'Person' },
+            { name: 'contents', type: 'string' },
+            { name: 'hash', type: 'bytes' },
+          ],
+        },
+        primaryType: 'Mail',
+        message: {
+          timestamp: 1234567890n,
+          contents: 'Hello, Bob! 🖤',
+          hash: '0x1234567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef',
+          from: {
+            name: {
+              first: 'Cow',
+              last: 'Burns',
+            },
+            wallet: '0xCD2a3d9F938E13CD947Ec05AbC7FE734Df8DD826',
+            age: 69,
+            favoriteColors: ['red', 'green', 'blue'],
+            isCool: false,
+          },
+          to: {
+            name: { first: 'Bob', last: 'Builder' },
+            wallet: '0xbBbBBBBbbBBBbbbBbbBbbbbBBbBbbbbBbBbbBBbB',
+            age: 70,
+            favoriteColors: ['orange', 'yellow', 'green'],
+            isCool: true,
+          },
+        },
       }),
     ).toBeDefined()
   })