+++ src/utils/encoding/toBytes.test.ts
@@ -55,7 +55,11 @@ describe('converts numbers to bytes', () => {
     `)
 
     expect(() => toBytes(-69)).toThrowErrorMatchingInlineSnapshot(
-      '"Number \\"-69\\" is not in safe integer range (0 to 9007199254740991)"',
+      `
+      "Number \\"-69\\" is not in safe integer range (0 to 9007199254740991)
+
+      Version: viem@1.0.2"
+    `,
     )
   })
 
@@ -83,17 +87,29 @@ describe('converts numbers to bytes', () => {
     expect(() =>
       numberToBytes(-7, { size: 1 }),
     ).toThrowErrorMatchingInlineSnapshot(
-      '"Number \\"-7\\" is not in safe 8-bit unsigned integer range (0 to 255)"',
+      `
+      "Number \\"-7\\" is not in safe 8-bit unsigned integer range (0 to 255)
+
+      Version: viem@1.0.2"
+    `,
     )
     expect(() =>
       numberToBytes(256, { size: 1 }),
     ).toThrowErrorMatchingInlineSnapshot(
-      '"Number \\"256\\" is not in safe 8-bit unsigned integer range (0 to 255)"',
+      `
+      "Number \\"256\\" is not in safe 8-bit unsigned integer range (0 to 255)
+
+      Version: viem@1.0.2"
+    `,
     )
     expect(() =>
       numberToBytes(65536, { size: 2 }),
     ).toThrowErrorMatchingInlineSnapshot(
-      '"Number \\"65536\\" is not in safe 16-bit unsigned integer range (0 to 65535)"',
+      `
+      "Number \\"65536\\" is not in safe 16-bit unsigned integer range (0 to 65535)
+
+      Version: viem@1.0.2"
+    `,
     )
   })
 
@@ -164,12 +180,20 @@ describe('converts numbers to bytes', () => {
     expect(() =>
       numberToBytes(32768, { size: 2, signed: true }),
     ).toThrowErrorMatchingInlineSnapshot(
-      '"Number \\"32768\\" is not in safe 16-bit signed integer range (-32768 to 32767)"',
+      `
+      "Number \\"32768\\" is not in safe 16-bit signed integer range (-32768 to 32767)
+
+      Version: viem@1.0.2"
+    `,
     )
     expect(() =>
       numberToBytes(-32769, { size: 2, signed: true }),
     ).toThrowErrorMatchingInlineSnapshot(
-      '"Number \\"-32769\\" is not in safe 16-bit signed integer range (-32768 to 32767)"',
+      `
+      "Number \\"-32769\\" is not in safe 16-bit signed integer range (-32768 to 32767)
+
+      Version: viem@1.0.2"
+    `,
     )
   })
 })
@@ -331,22 +355,38 @@ describe('converts bigints to bytes', () => {
     expect(() =>
       numberToBytes(-7n, { size: 1 }),
     ).toThrowErrorMatchingInlineSnapshot(
-      '"Number \\"-7n\\" is not in safe 8-bit unsigned integer range (0n to 255n)"',
+      `
+      "Number \\"-7n\\" is not in safe 8-bit unsigned integer range (0n to 255n)
+
+      Version: viem@1.0.2"
+    `,
     )
     expect(() =>
       numberToBytes(256n, { size: 1 }),
     ).toThrowErrorMatchingInlineSnapshot(
-      '"Number \\"256n\\" is not in safe 8-bit unsigned integer range (0n to 255n)"',
+      `
+      "Number \\"256n\\" is not in safe 8-bit unsigned integer range (0n to 255n)
+
+      Version: viem@1.0.2"
+    `,
     )
     expect(() =>
       numberToBytes(65536n, { size: 2 }),
     ).toThrowErrorMatchingInlineSnapshot(
-      '"Number \\"65536n\\" is not in safe 16-bit unsigned integer range (0n to 65535n)"',
+      `
+      "Number \\"65536n\\" is not in safe 16-bit unsigned integer range (0n to 65535n)
+
+      Version: viem@1.0.2"
+    `,
     )
     expect(() =>
       numberToBytes(18446744073709551616n, { size: 8 }),
     ).toThrowErrorMatchingInlineSnapshot(
-      '"Number \\"18446744073709551616n\\" is not in safe 64-bit unsigned integer range (0n to 18446744073709551615n)"',
+      `
+      "Number \\"18446744073709551616n\\" is not in safe 64-bit unsigned integer range (0n to 18446744073709551615n)
+
+      Version: viem@1.0.2"
+    `,
     )
   })
 
@@ -502,7 +542,11 @@ describe('converts bigints to bytes', () => {
         signed: true,
       }),
     ).toThrowErrorMatchingInlineSnapshot(
-      '"Number \\"170141183460469231731687303715884105728n\\" is not in safe 128-bit signed integer range (-170141183460469231731687303715884105728n to 170141183460469231731687303715884105727n)"',
+      `
+      "Number \\"170141183460469231731687303715884105728n\\" is not in safe 128-bit signed integer range (-170141183460469231731687303715884105728n to 170141183460469231731687303715884105727n)
+
+      Version: viem@1.0.2"
+    `,
     )
     expect(() =>
       numberToBytes(-170141183460469231731687303715884105729n, {
@@ -510,7 +554,11 @@ describe('converts bigints to bytes', () => {
         signed: true,
       }),
     ).toThrowErrorMatchingInlineSnapshot(
-      '"Number \\"-170141183460469231731687303715884105729n\\" is not in safe 128-bit signed integer range (-170141183460469231731687303715884105728n to 170141183460469231731687303715884105727n)"',
+      `
+      "Number \\"-170141183460469231731687303715884105729n\\" is not in safe 128-bit signed integer range (-170141183460469231731687303715884105728n to 170141183460469231731687303715884105727n)
+
+      Version: viem@1.0.2"
+    `,
     )
   })
 })