+++ src/utils/index.ts
@@ -70,6 +70,9 @@ export {
 } from './contract'
 
 export {
+  concat,
+  concatBytes,
+  concatHex,
   isBytes,
   isHex,
   pad,