+++ src/utils/transaction/prepareRequest.test.ts
@@ -11,7 +11,7 @@ import { parseEther, parseGwei } from '../../utils'
 import * as publicActions from '../../actions/public'
 
 import { defaultTip, prepareRequest } from './prepareRequest'
-import { getEoaAccount } from '../../_test/utils'
+import { getLocalAccount } from '../../_test/utils'
 
 const sourceAccount = accounts[0]
 const targetAccount = accounts[1]
@@ -43,7 +43,7 @@ describe('prepareRequest', () => {
       nonce: _nonce,
       ...rest
     } = await prepareRequest(walletClient, {
-      account: getEoaAccount(sourceAccount.privateKey),
+      account: getLocalAccount(sourceAccount.privateKey),
       to: targetAccount.address,
       value: parseEther('1'),
     })
@@ -57,6 +57,7 @@ describe('prepareRequest', () => {
           "address": "0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266",
           "signMessage": [Function],
           "signTransaction": [Function],
+          "signTypedData": [Function],
           "type": "local",
         },
         "from": "0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266",
@@ -76,7 +77,7 @@ describe('prepareRequest', () => {
     } as any)
 
     const { nonce: _nonce, ...request } = await prepareRequest(walletClient, {
-      account: getEoaAccount(sourceAccount.privateKey),
+      account: getLocalAccount(sourceAccount.privateKey),
       to: targetAccount.address,
       value: parseEther('1'),
     })
@@ -86,6 +87,7 @@ describe('prepareRequest', () => {
           "address": "0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266",
           "signMessage": [Function],
           "signTransaction": [Function],
+          "signTypedData": [Function],
           "type": "local",
         },
         "from": "0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266",
@@ -105,7 +107,7 @@ describe('prepareRequest', () => {
       nonce: _nonce,
       ...rest
     } = await prepareRequest(walletClient, {
-      account: getEoaAccount(sourceAccount.privateKey),
+      account: getLocalAccount(sourceAccount.privateKey),
       to: targetAccount.address,
       value: parseEther('1'),
     })
@@ -115,6 +117,7 @@ describe('prepareRequest', () => {
           "address": "0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266",
           "signMessage": [Function],
           "signTransaction": [Function],
+          "signTypedData": [Function],
           "type": "local",
         },
         "from": "0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266",
@@ -132,7 +135,7 @@ describe('prepareRequest', () => {
     const { maxFeePerGas: _maxFeePerGas, ...rest } = await prepareRequest(
       walletClient,
       {
-        account: getEoaAccount(sourceAccount.privateKey),
+        account: getLocalAccount(sourceAccount.privateKey),
         to: targetAccount.address,
         nonce: 5,
         value: parseEther('1'),
@@ -144,6 +147,7 @@ describe('prepareRequest', () => {
           "address": "0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266",
           "signMessage": [Function],
           "signTransaction": [Function],
+          "signTypedData": [Function],
           "type": "local",
         },
         "from": "0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266",
@@ -162,7 +166,7 @@ describe('prepareRequest', () => {
     vi.spyOn(publicActions, 'getBlock').mockResolvedValueOnce({} as any)
 
     const { nonce: _nonce, ...request } = await prepareRequest(walletClient, {
-      account: getEoaAccount(sourceAccount.privateKey),
+      account: getLocalAccount(sourceAccount.privateKey),
       to: targetAccount.address,
       gasPrice: parseGwei('10'),
       value: parseEther('1'),
@@ -173,6 +177,7 @@ describe('prepareRequest', () => {
           "address": "0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266",
           "signMessage": [Function],
           "signTransaction": [Function],
+          "signTypedData": [Function],
           "type": "local",
         },
         "from": "0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266",
@@ -189,7 +194,7 @@ describe('prepareRequest', () => {
 
     await expect(() =>
       prepareRequest(walletClient, {
-        account: getEoaAccount(sourceAccount.privateKey),
+        account: getLocalAccount(sourceAccount.privateKey),
         to: targetAccount.address,
         gasPrice: parseGwei('10'),
         value: parseEther('1'),
@@ -205,7 +210,7 @@ describe('prepareRequest', () => {
     await setup()
 
     const { nonce: _nonce, ...rest } = await prepareRequest(walletClient, {
-      account: getEoaAccount(sourceAccount.privateKey),
+      account: getLocalAccount(sourceAccount.privateKey),
       to: targetAccount.address,
       maxFeePerGas: parseGwei('10'),
       value: parseEther('1'),
@@ -216,6 +221,7 @@ describe('prepareRequest', () => {
           "address": "0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266",
           "signMessage": [Function],
           "signTransaction": [Function],
+          "signTypedData": [Function],
           "type": "local",
         },
         "from": "0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266",
@@ -233,7 +239,7 @@ describe('prepareRequest', () => {
 
     await expect(() =>
       prepareRequest(walletClient, {
-        account: getEoaAccount(sourceAccount.privateKey),
+        account: getLocalAccount(sourceAccount.privateKey),
         to: targetAccount.address,
         maxFeePerGas: parseGwei('1'),
         value: parseEther('1'),
@@ -254,7 +260,7 @@ describe('prepareRequest', () => {
 
     await expect(() =>
       prepareRequest(walletClient, {
-        account: getEoaAccount(sourceAccount.privateKey),
+        account: getLocalAccount(sourceAccount.privateKey),
         to: targetAccount.address,
         maxFeePerGas: parseGwei('10'),
         value: parseEther('1'),
@@ -270,7 +276,7 @@ describe('prepareRequest', () => {
     await setup()
 
     const { nonce: _nonce, ...rest } = await prepareRequest(walletClient, {
-      account: getEoaAccount(sourceAccount.privateKey),
+      account: getLocalAccount(sourceAccount.privateKey),
       to: targetAccount.address,
       maxPriorityFeePerGas: parseGwei('5'),
       value: parseEther('1'),
@@ -281,6 +287,7 @@ describe('prepareRequest', () => {
           "address": "0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266",
           "signMessage": [Function],
           "signTransaction": [Function],
+          "signTypedData": [Function],
           "type": "local",
         },
         "from": "0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266",
@@ -302,7 +309,7 @@ describe('prepareRequest', () => {
 
     await expect(() =>
       prepareRequest(walletClient, {
-        account: getEoaAccount(sourceAccount.privateKey),
+        account: getLocalAccount(sourceAccount.privateKey),
         to: targetAccount.address,
         maxFeePerGas: parseGwei('5'),
         value: parseEther('1'),
@@ -318,7 +325,7 @@ describe('prepareRequest', () => {
     await setup()
 
     const { nonce: _nonce, ...rest } = await prepareRequest(walletClient, {
-      account: getEoaAccount(sourceAccount.privateKey),
+      account: getLocalAccount(sourceAccount.privateKey),
       to: targetAccount.address,
       maxFeePerGas: parseGwei('10'),
       maxPriorityFeePerGas: parseGwei('5'),
@@ -330,6 +337,7 @@ describe('prepareRequest', () => {
           "address": "0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266",
           "signMessage": [Function],
           "signTransaction": [Function],
+          "signTypedData": [Function],
           "type": "local",
         },
         "from": "0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266",